campaign-backend/requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
certifi==2025.1.31
click==8.1.8
fastapi==0.115.7
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
mako==1.3.8
markupsafe==3.0.2
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-settings==2.7.1
pytest==8.3.4
pytest-asyncio==0.25.3
python-decouple==3.8
python-dotenv==1.0.1
sniffio==1.3.1
sqlalchemy==2.0.37
starlette==0.45.3
typing-extensions==4.12.2
uvicorn==0.34.0
